refactoring and cleaning up.
